app/api/requirements.txt
faiss-cpu==1.7.4
flask==2.3.2 
numpy==1.24.3
pandas==1.5.3
sentence-transformers==2.2.2